+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Documentation:
 
 * Updated the badges ion the README to be a little more on brand and relevant. ([@TGWolf][])
 
+New Features:
+
+* Added a new script verbose-mode/verbose-mode.sh. ([@TGWolf][])
+
 ## [v1.0.8] - December 11 2019
 
 New Features: